feat(vendingmachine): HoldingSum, Stock 객체 예외 처리 추가
