+++ src/main/java/vendingmachine/model/HoldingSum.java
@@ -5,6 +5,7 @@ import java.util.Map;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import vendingmachine.util.NumberChecker;
+import vendingmachine.util.StringChecker;
 
 public class HoldingSum {
 
@@ -20,6 +21,9 @@ public class HoldingSum {
 	}
 
 	private void checkInput(String input) {
+		StringChecker stringChecker = new StringChecker();
+		stringChecker.isEmpty(input);
+
 		NumberChecker numberChecker = new NumberChecker();
 		numberChecker.isPositiveInteger(input);
 		numberChecker.isDivisibleNumber(input, Coin.minAmount());
